requirements.txt
aiofiles==0.8.0
ajsonrpc==1.2.0
anyio==3.5.0
appdirs==1.4.4
argcomplete==1.12.3
arrow==1.2.2
asgiref==3.5.0
Beaker==1.10.0
beautifulsoup4==4.9.3
blivet==3.4.3
blivet-gui==2.3.0
bottle==0.12.19
Brlapi==0.8.2
bs4==0.0.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
chrome-gnome-shell==0.0.0
click==8.0.4
colorama==0.4.4
cryptography==3.4.7
cupshelpers==1.0
dasbus==1.6
dbus-python==1.2.18
decorator==5.0.9
distro==1.6.0
docopt==0.6.2
fedora-third-party==0.8
fros==1.1
gpg==1.15.1
greenlet==1.1.1
gssapi==1.6.14
h11==0.13.0
humanize==0.5.1
idna==3.2
ifaddr==0.1.7
image-go-nord==0.1.4
koji==1.28.0
langtable==0.0.57
libcomps==0.1.18
libsixel-python==0.5.0
lxml==4.6.5
Mako==1.1.4
MarkupSafe==2.0.0
marshmallow==3.14.1
msgpack==1.0.3
nftables==0.1
notify2==0.3.1
numpy==1.21.5
olefile==0.46
opencv-python==4.5.5.62
packaging==21.3
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==8.3.2
pipreqs==0.4.11
platformio==5.2.5
ply==3.11
pocket==0.3.6
productmd==1.33
progress==1.6
progressbar2==3.53.2
ptyprocess==0.6.0
pwquality==1.4.4
pycairo==1.20.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.28
pyenchant==3.2.2
pyfiglet==0.8.post1
pygit2==1.6.1
PyGObject==3.42.0
pykickstart==3.34
pynvim==0.4.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.7
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.9
python-augeas==1.1.0
python-dateutil==2.8.1
python-meh==0.50
python-resize-image==1.1.20
python-utils==2.5.6
pyudev==0.22.0
pyxdg==0.27
rainbowstream==1.6.0
regex==2022.1.18
requests==2.27.0
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
resize-image==0.4.0
rpm==4.17.0
rpmautospec==0.2.5
scour==0.38.1
selinux==3.3
semantic-version==2.9.0
sepolicy==3.3
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.8.2
six==1.16.0
sniffio==1.2.0
sos==4.2
soupsieve==2.3.1
starlette==0.18.0
systemd-python==234
tabulate==0.8.9
Tempita==0.5.2
twitter==1.19.3
urllib3==1.26.7
uvicorn==0.17.5
wsproto==1.0.0
yarg==0.1.9
zeroconf==0.38.4
